LayoutTests/fast/events/touch/ios/touch-event-radius-expected.txt

PASS: radiusX is correct.
PASS: radiusY is correct.
